# API/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API del laboratorio de IoT",
    "description": "API para la gestión de dispositivos, modelos y sensores. \n \n Cuando se quiere insertar un dato no hay que poner el campo \"_id\", ya que es autogenerado por la base de datos.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/records/{cameraId}": {
      "get": {
        "tags": [
          "Records"
        ],
        "summary": "Obtener todas las grabaciones",
        "description": "Retorna una lista de todas las grabaciones según el filtro.",
        "parameters": [
          {
            "in": "path",
            "name": "cameraId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la cámara"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de grabaciones obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Record"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener las grabaciones"
          }
        }
      }
    },
...
    "/record/{id}": {
      "get": {
        "tags": [
          "Records"
        ],
        "summary": "Obtener una grabación por ID",
        "description": "Retorna una grabación por su ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la grabación"
          }
        ],
        "responses": {
          "200": {
            "description": "Grabación obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener la grabación"
          }
        }
      },
      "put": {
        "tags": [
          "Records"
        ],
        "summary": "Actualizar una grabación por ID",
        "description": "Actualiza la información de una grabación específica.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la grabación"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Record"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Grabación actualizada correctamente"
          },
          "500": {
            "description": "Error al actualizar la grabación"
          }
        }
      },
      "delete": {
        "tags": [
          "Records"
        ],
        "summary": "Eliminar una grabación por ID",
        "description": "Elimina una grabación específica por su ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la grabación"
          }
        ],
        "responses": {
          "200": {
            "description": "Grabación eliminada correctamente"
          },
          "500": {
            "description": "Error al eliminar la grabación"
          }
        }
      }
    },
    "/roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Obtener todos los roles",
        "description": "Retorna una lista de todos los roles.",
        "responses": {
          "200": {
            "description": "Lista de roles obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener los roles"
          }
        }
      },
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "Actualizar un rol por ID",
        "description": "Actualiza la información de un rol específico.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del rol"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rol actualizado correctamente"
          },
          "500": {
            "description": "Error al actualizar el rol"
          }
        }
      },
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "Crear un nuevo rol",
        "description": "Crea un nuevo rol.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rol creado correctamente"
          },
          "500": {
            "description": "Error al crear el rol"
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "summary": "Eliminar un rol por ID",
        "description": "Elimina un rol específico por su ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del rol"
          }
        ],
        "responses": {
          "200": {
            "description": "Rol eliminado correctamente"
          },
          "500": {
            "description": "Error al eliminar el rol"
          }
        }
      }
    },
    "/roles/{id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Obtener un rol por ID",
        "description": "Retorna un rol por su ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del rol"
          }
        ],
        "responses": {
          "200": {
            "description": "Rol obtenido correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener el rol"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obtener todos los usuarios",
        "description": "Retorna una lista de todos los usuarios.",
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener los usuarios"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Crear un nuevo usuario",
        "description": "Crea un nuevo usuario.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario creado correctamente"
          },
          "500": {
            "description": "Error al crear el usuario"
          }
        }
      }
    },
    "/users/signin": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Iniciar sesión",
        "description": "Inicia sesión en el sistema.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Correo electrónico del usuario",
                    "required": true
                  },
                  "password": {
                    "type": "string",
                    "description": "Contraseña del usuario",
                    "required": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión correcto"
          },
          "500": {
            "description": "Error al iniciar sesión"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obtener un usuario por ID",
        "description": "Retorna un usuario por su ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario obtenido correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener el usuario"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Permission": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID del permiso"
          },
          "module": {
            "type": "string",
            "description": "Módulo del permiso"
          },
          "action": {
            "type": "object",
            "properties": {
              "read": {
                "type": "boolean",
                "default": false
              },
              "use": {
                "type": "boolean",
                "default": false
              },
              "modify": {
                "type": "boolean",
                "default": false
              }
            }
          }
        }
      },
      "Record": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID de la grabación"
          },
          "device": {
            "type": "string",
            "description": "Dispositivo que graba"
          },
          "format": {
            "type": "string",
            "description": "Formato de la grabación"
          },
          "duration": {
            "type": "number",
            "description": "Duración de la grabación"
          },
          "filesize": {
            "type": "number",
            "description": "Tamaño del archivo de la grabación"
          },
          "fileUrl": {
            "type": "string",
            "description": "URL del archivo de la grabación"
          },
          "statTime": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de inicio de la grabación"
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de fin de la grabación"
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID del rol"
          },
          "name": {
            "type": "string",
            "description": "Nombre del rol"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "description": "Permisos del rol"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID del usuario"
          },
          "username": {
            "type": "string",
            "required": true,
            "unique": true,
            "description": "Nombre del usuario"
          },
          "password": {
            "type": "string",
            "required": true,
            "description": "Contraseña del usuario"
          },
          "fullname": {
            "type": "string",
            "description": "Nombre completo del usuario"
          },
          "email": {
            "type": "string",
            "required": true,
            "description": "Correo electrónico del usuario"
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "Rol del usuario"
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "description": "Último inicio de sesión"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación del usuario"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}